// convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  collections: typeof collections;
  crons: typeof crons;
  files: typeof files;
  http: typeof http;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;